Move GlobalShortcutManager to its own file
